--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -316,6 +316,9 @@ endif()
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src/External)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 
+option(USE_PRECOMPILED_HEADERS
+  "Use precompiled headers to speed up compilation" ON)
+
 add_subdirectory(src/Cello)
 add_subdirectory(src/Enzo)
 add_subdirectory(src/External)